requirements.txt
Django==4.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
psycopg2-binary==2.9.6
docker==5.0.3
pytest==7.2.2
pytest-django==4.5.2
swagger-ui-bundle==0.0.8
drf-yasg